+++ .eslintrc.js
@@ -31,7 +31,8 @@ module.exports = {
         '@typescript-eslint/no-explicit-any': 0,
         '@typescript-eslint/no-non-null-assertion': 'off',
         'react-hooks/rules-of-hooks': 'error',
-        'react-hooks/exhaustive-deps': 'warn'
+        'react-hooks/exhaustive-deps': 'warn',
+        'no-param-reassign': ['error', { props: true, ignorePropertyModificationsFor: ['state'] }]
     },
     settings: {
         react: {
