+++ .prettierrc
@@ -1,5 +1,4 @@
 {
-    "printWidth": 160,
     "tabWidth": 4,
     "singleQuote": true,
     "semi": false,
