+++ package.json
@@ -7,6 +7,7 @@
         "@emotion/styled": "^11.10.4",
         "@mui/icons-material": "^5.10.3",
         "@mui/material": "^5.10.4",
+        "@reduxjs/toolkit": "^1.8.5",
         "@testing-library/jest-dom": "^5.11.4",
         "@testing-library/react": "^13.1.1",
         "@testing-library/user-event": "^14.1.1",
@@ -25,7 +26,6 @@
         "react-router-dom": "^6.1.1",
         "react-scripts": "5.0.1",
         "redux-persist": "^6.0.0",
-        "redux-thunk": "^2.4.1",
         "sass": "^1.52.1",
         "sass-loader": "^13.0.0",
         "typescript": "^4.7.2"
