+++ src/App.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import { Provider } from 'react-redux'
 import { PersistGate } from 'redux-persist/integration/react'
 import RootComponent from './RootComponent'
-import { persistor, store } from './store/reducers/store'
+import { persistor, store } from './redux/store'
 
 const App: React.FC = () => {
     return (
