+++ src/redux/store.ts
@@ -1,0 +1,30 @@
+import { Action, configureStore, ThunkAction } from '@reduxjs/toolkit'
+import { combineReducers } from 'redux'
+import { FLUSH, PAUSE, PERSIST, persistReducer, persistStore, PURGE, REGISTER, REHYDRATE } from 'redux-persist'
+import storage from 'redux-persist/lib/storage' // defaults to localStorage
+import data from './reducers/data'
+
+const persistConfig = {
+    key: 'root',
+    storage,
+    whitelist: ['data'], // elements that will be persisted
+    blacklist: [] // elements that will not be persisted
+}
+
+const rootReducer = combineReducers({
+    data
+})
+
+const persistedReducer = persistReducer(persistConfig, rootReducer)
+
+const store = configureStore({
+    reducer: persistedReducer,
+    middleware: (getDefaultMiddleware) => getDefaultMiddleware({ serializableCheck: { ignoredActions: [FLUSH, REHYDRATE, PAUSE, PERSIST, PURGE, REGISTER] } })
+})
+const persistor = persistStore(store)
+
+export { store, persistor }
+
+export type RootState = ReturnType<typeof store.getState>
+export type AppDispatch = typeof store.dispatch
+export type AppThunk<ReturnType = void> = ThunkAction<ReturnType, RootState, unknown, Action<string>>
