--- src/utility/functions.ts
@@ -1,8 +1,0 @@
-/**
- * This function can be used anywhere in the app to greet the user
- * @param userName The user's first name
- * @returns A kind greeting message
- */
-export const sayHello = (userName: string): string => {
-    return 'Welcome ' + userName + '!'
-}
